Gradients are tested and ready to merge.
Things were almost right, wrote two tests, one for `d>1` and the
other for `d>1` and multiple test points. Both use a `braninxy.npz`
dataset that was generated observing the noiseless functions at 30
uniform random points.

Note: the dataset could in principle be generated by the test, thus
eliminating the need for an ugly dataset in the test directory but
that requires hard coding the branin function (or another 2d
function) into the file; not that big a deal, I just haven't done it
(yet).