Added feature to log likelihood.
gp.loglikelihood() now text an optional `n` keyword argument. If
`n < gp.ndata`, the function returns the log likelihood of the last
`n` observations conditioned on all the previous.

Note: the gradient for `n < gp.ndata` will not return the correct
gradient, in fact it will not even be the right size. This still
needs to be fixed.