SMC seems to be working.
Running demos/sampling.py now outputs an extra figure corresponding
to the Bayesian posterior using 200 SMC samples. I gave each
particle 10 burn-in samples to give them a chance to settle into
the posterior. In practice, the data would presumably be added
sequentially and the posterior would change slowly enough so that
a single MCMC step between data points should be good enough.

The particular numbers 200 and 10 were simply chosen to roughly
correspond to the amount of work done by the MCMC method (2000
samples with a burn-in of 100).